requirements.txt
python-telegram-bot==21.0
vk-api==11.9.9
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
